requirements.txt
APScheduler==3.6.3
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
cachetools==4.2.2
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
cycler==0.10.0
distlib==0.3.3
docutils==0.18
et-xmlfile==1.1.0
filelock==3.4.0
Flask==2.0.2
imageio==2.11.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
opencv-contrib-python==4.5.4.58
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
pipenv==2021.11.15
platformdirs==2.4.0
pycodestyle==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
statistics==1.0.3.5
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
virtualenv==20.10.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
XlsxWriter==3.0.2
